# swagger.yaml
openapi: 3.0.0
info:
  title: Bookstore API
  version: 1.0.1
description: |
  REST API for managing a bookstore.

paths:
  /customer/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: User registered successfully
        "400":
          description: Invalid request

  /books:
    get:
      summary: Get the list of available books
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Books"

  /isbn/{isbn}:
    get:
      summary: Get book details based on ISBN
      parameters:
        - name: isbn
          in: path
          required: true
          description: ISBN of the book
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

  /author/{author}:
    get:
      summary: Get books based on author
      parameters:
        - name: author
          in: path
          required: true
          description: Author name
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BooksByAuthor"

  /title/{title}:
    get:
      summary: Get books based on title
      parameters:
        - name: title
          in: path
          required: true
          description: Book title
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BooksByTitle"

  /review/{isbn}:
    get:
      summary: Get book review
      parameters:
        - name: isbn
          in: path
          required: true
          description: ISBN of the book
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookReviews"

  /customer/login:
    post:
      summary: User Login
      description: Logs in a registered user and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /customer/auth/review/{isbn}:
    put:
      summary: Add or Update Book Review
      description: Adds or updates a book review for the authenticated user.
      parameters:
        - name: isbn
          in: path
          required: true
          description: ISBN of the book
          schema:
            type: string
        - name: review
          in: query
          required: true
          description: Reviewer's comment
          schema:
            type: string
      responses:
        "200":
          description: Review updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

    delete:
      summary: Delete Book Review
      description: Deletes the review of the authenticated user for the specified book.
      parameters:
        - name: isbn
          in: path
          required: true
          description: ISBN of the book
          schema:
            type: string
      responses:
        "200":
          description: Review deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string

    Book:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        reviews:
          type: object

    Books:
      type: array
      items:
        $ref: "#/components/schemas/Book"

    BooksByAuthor:
      type: object
      properties:
        booksbyauthor:
          $ref: "#/components/schemas/Books"

    BooksByTitle:
      type: object
      properties:
        booksbytitle:
          $ref: "#/components/schemas/Books"

    BookReviews:
      type: object
      additionalProperties:
        type: string